# dev(m_engine): switch to lexer with macro engine