--- a/crates/mitex-parser/src/lib.rs
+++ b/crates/mitex-parser/src/lib.rs
@@ -14,10 +14,10 @@ use parser::Parser;
 ///
 /// The error nodes are attached to the tree
 pub fn parse(input: &str, spec: CommandSpec) -> SyntaxNode {
-    SyntaxNode::new_root(Parser::new(input, spec).parse())
+    SyntaxNode::new_root(Parser::new_macro(input, spec).parse())
 }
 
 /// It is only for internal testing
-pub fn parse_with_macro(input: &str, spec: CommandSpec) -> SyntaxNode {
-    SyntaxNode::new_root(Parser::new_macro(input, spec).parse())
+pub fn parse_without_macro(input: &str, spec: CommandSpec) -> SyntaxNode {
+    SyntaxNode::new_root(Parser::new(input, spec).parse())
 }